Make sure we build out of tree where possible
Even if we're dealing with autotools based projects, we should still
do an out of tree build, otherwise we'll run into trouble when switching
between distros as make won't know it has to rebuild everything when e.g.
glibc changes.

To implement this, if the project has a working "make install", we just
copy all files to the build directory and do the build from the build
directory. The build directory is separate per distribution which means
this works fine.

For xfstests, which does not have a working make install, we're forced
to run make clean every time instead.