Build most modules as external modules instead of builtin modules
Building as external modules will allow us to provide these via a
virtiofs mount and which means we can update them while the image
is booted.

This does imply we need an initramfs now so we update the configuration
to make sure one is built by default. Additionally, we prevent any
bootloader from getting installed instead as we still do direct kernel
boot and thus don't have any need for an ESP or bootloader by default.