Use overlayfs for fstests as well
For fstests, we use overlayfs in two ways. The first is to make sure
the build outputs are separate from the sources like we use it for
blktests and btrfs-progs.

The second way is to hack around the limitations of make install for
fstests, so that we can now use make install for fstests as well.