Don't use initrd by default
Let's build an ext4 image and let's not make it bootable by default.
This allows us to direct kernel boot without an initrd.

We make ext4 builtin to make this possible, since I'm not aware of
anyone hacking on ext4 whereas I do know people hacking on btrfs with
mkosi-kernel.