profiles: add cxl profile
This profile adds a 1GB CXL memory expander to the QemuArgs, along with
the MaxMem and CXL runtime arguments for mkosi.

A simple systemd service is added to automatically map the cxl device as
system-ram.

This profile requires that your kernel be built with:
	CONFIG_CXL_REGION_INVALIDATION_TEST=y

This is off by default.  It bypasses a cache invalidation warning that
otherwise would cause the CXL device to fail to probe.

Signed-off-by: Gregory Price <gourry@gourry.net>